게시글 좋아요 완료

+++ Justsurviveoffice/src/main/java/com/spring/app/comment/domain/CommentDTO.java
@@ -32,7 +32,7 @@ public class CommentDTO {
 
 	private Long fk_boardNo;
 
-	private int parentNo;
+	private Long parentNo;
 	
 	private Users users;
